// frontend/src/api/models/AddOnUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AddOnCategory } from './AddOnCategory';
export type AddOnUpdate = {
    code?: (string | null);
    name?: (string | null);
    description?: (string | null);
    price?: (number | string | null);
    category?: (AddOnCategory | null);
    is_active?: (boolean | null);
};